backend/pyproject.toml
[project]
name = "second-brain"
version = "0.1.0"
description = "Personal Knowledge Base with AI-Powered Search"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "pgvector>=0.3.6",
    "httpx>=0.28.0",
    "python-multipart>=0.0.17",
    # AI/ML
    "openai>=1.57.0",
    "anthropic>=0.39.0",
    "tiktoken>=0.8.0",
    # Content processing
    "pdfplumber>=0.11.4",
    "trafilatura>=1.14.0",
    "readability-lxml>=0.8.1",
    "beautifulsoup4>=4.12.3",
    # Telegram
    "python-telegram-bot[webhooks]>=22.0",
    # Utilities
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "factory-boy>=3.3.1",
    "aiosqlite>=0.20.0",
]
ocr = [
    "pytesseract>=0.3.13",
    "pdf2image>=1.17.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = ["ignore::DeprecationWarning"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.ruff.lint.isort]
known-first-party = ["app", "telegram_bot"]
